First Commit
This is the website for converting hand drawn "Images to Html code" using "Deep Neural Technology". This website mainly consist of two page, one is "Home Page" container input box for uploading the image and another another text box containing the "Generated Html Code" and buttons for Generate Code and download the code and also a progress bar showing the steps to be applied while using this website to acheive  your required output. There is also a text box where you can the enter the name of the file and its extension in which you want to download the Html Code. Second Page are "About Page" containing the description of the Project and some instruction for the user who want to use this tool and in last there is a contect portion through you are in touch with us.